\name{catPkg2-package}
\alias{catPkg2-package}
\alias{catPkg2}
\docType{package}
\title{
\packageTitle{catPkg2}
}
\description{
\packageDescription{catPkg2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{catPkg2}
\packageIndices{catPkg2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{catPkg2}

Maintainer: \packageMaintainer{catPkg2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
