% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Clean_String.R
\name{Clean_String}
\alias{Clean_String}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
Clean_String(str)
}
\arguments{
\item{str}{A single input string such as "This is a cool function!"}
}
\value{
A vector contain all valid tokens in the original input string
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}

